Platform/O6: Fix rebuild of csupm_bin_config
csupm_bin_config is a build target which automatically rebuilds the
output file, but the output file is not rebuilt upon second call if
this binary is not removed, and that file is also moved away so this
results in a build error. Let's designate the output file as the target
instead and csupm_bin_config as its dependency.

Signed-off-by: Willy Tarreau <w@1wt.eu>