Platform/O6: Add pm_config to permit configuration of OPP
This contains the definitions of the OPP that can be adjusted
directly in the code.

It was copied from edk2-non-osi/Platform/CIX/Sky1/PackageTool/pm_config.

This requires edk2-non-osi with the following commit:

  commit 7ab8325ec818743dd62d19765d7ffdbeebc27bfc
  Author: ReleaseRaptor <ReleaseRaptor@cixtech.com>
  Date:   Wed Feb 26 22:39:27 2025 +0800

      CIX Firmware Auto Release

      Update to P1 Beta2 RELEASE build of commit 0a8258da4512

      Change-Id: Ia129cec25a4f57418e0f05836f23306e37f544d5

But not any newer one as the init code uses different pm format
and is no longer compatible.

Signed-off-by: Willy Tarreau <w@1wt.eu>