WIP: set CPU frequencies to 3.0+2.7+2.6+2.6
According to some CIX files, limits should be 2.8 for both big,
2.6 for both med and 2.0 for both little. The default settings
are:
  b0 2.4
  b1 2.5
  m0 2.3
  m1 2.2

Let's increase the voltage for b0 and m1 compared to b1 and m0
to compensate for the extra effort. My machine now works pretty
solid during several hours builds with this:

  b0: 2.7 (it seldom crashes at 2.8, even with voltage up to 1050mV,
           while it's rock solid at 2.7/950mV for hours)
  b1: 3.0 at 970mV
  m0: 2.6 at 970mV
  m1: 2.6 at 990mV

The total capacity of the A720 cores is now 9% above the stock 2.6+2.4,
and the peak is 15% above.