Platform/O6: set CPU frequencies to 3.0+2.7+2.6+2.6
According to some CIX files, hard limits should be 2.8 GHz for both
big, 2.6 for both med and 2.0 for both little. The default settings
are:
  b0: 2.4 GHz (cpu 9-10)
  b1: 2.5 GHz (cpu 0,11)
  m0: 2.3 GHz (cpu 5-6)
  m1: 2.2 GHz (cpu 7-8)

Let's increase the voltage for b0 and m1 compared to b1 and m0
to compensate for the extra effort. My machine now works pretty
solid during several hours builds with this:

  b0: 2.7 (it seldom crashes at 2.8, even with voltage up to 1050mV,
           while it's rock solid at 2.7/950mV for hours)
  b1: 3.0 at 970mV
  m0: 2.6 at 970mV
  m1: 2.6 at 990mV

The total capacity of the A720 cores is now 9% above the advertised
2.6+2.4, 13.5% above the stock settings, the peak is 15.4% above the
original one and the worst case 18.2% above.

Signed-off-by: Willy Tarreau <w@1wt.eu>