Platform/O6: Add opp for 2x2.6 GHz + 2x2.4 GHz
The two big clusters now run at 2.6 GHz under 920 mV, the med ones
at 2.4 GHz under 920 mV. The little ones are unchanged (1.8 GHz under
790 mV). This is supposed to work according to the public specs, and
for now works on two tested chips which still show a significant margin
above that.

This provides a 4.2% avg perf increase of the stock 2.6+2.5+2.3+2.2,
and a 9.1% gain for the worst case.

Signed-off-by: Willy Tarreau <w@1wt.eu>